using service points to distinguish the door/window

+++ api/internal/types/queue.go
@@ -8,7 +8,7 @@ type Entry struct {
 	TicketNumber  string    `bson:"ticketNumber" json:"ticketNumber"`
 	QRToken       string    `bson:"qrToken" json:"qrToken"`
 	Status        string    `bson:"status" json:"status"` // WAITING, CALLED, IN_SERVICE, COMPLETED, SKIPPED, CANCELLED, NO_SHOW
-	Position      int       `bson:"position" json:"position"`
+	Position      int64     `bson:"position" json:"position"`
 	ServicePoint  string    `bson:"servicePoint,omitempty" json:"servicePoint,omitempty"` // Which service point (door/window) to go to
 	CreatedAt     time.Time `bson:"createdAt" json:"createdAt"`
 	UpdatedAt     time.Time `bson:"updatedAt" json:"updatedAt"`
